Refactor: fix imports after moving pluginManager to ilastik.plugins.manager
also removed import guards - ilastik (object classification) is pretty useless without, and
I don't see much of a reason to not fail here.
