+++ tests/test_ilastik/test_applets/objectExtraction/testOperators.py
@@ -29,7 +29,7 @@ import vigra
 from lazyflow.graph import Graph
 from lazyflow.operators import OpLabelVolume
 from ilastik.applets.objectExtraction.opObjectExtraction import OpAdaptTimeListRoi, OpRegionFeatures, OpObjectExtraction
-from ilastik.plugins import pluginManager
+from ilastik.plugins.manager import pluginManager
 
 import warnings
